#285: split DashboardController into resource-per-controller files
Pure refactor — 1700-line monolith → ~19 smaller controllers under
app/Http/Controllers/Dashboard/, each owning a single dashboard
resource. Shared env-loading + oauth-row-shaping helpers move to a
ResolvesDashboardEnv trait. routes/dashboard.php updated to point
each route at its new controller; behavior unchanged.
